app/requirements.txt
# Basic
Django>=3.0,<4.0
gunicorn==20.1.0
psycopg2-binary==2.9.1

# Other
django-tinymce==3.4.0

# Tests
coverage>=6.0,<7.0
pytest-django>=4.0,<5.0

# API REST
djangorestframework>=3.13.1<=3.14
